requirement.txt
Django==1.8.4
Pillow==3.1.0
sorl-thumbnail==12.3
wheel==0.24.0
